+++ src/variables/themes/index.js
@@ -1,0 +1,6 @@
+import darkTheme from './darkTheme'
+import lightTheme from './lightTheme'
+export { 
+	darkTheme,
+	lightTheme
+}
