+++ src/views/Projects/ProjectCards/ProjectData.js
@@ -188,7 +188,6 @@ class ProjectData extends PureComponent {
 	renderMenu = () => {
 		const { actionAnchor, actionAnchorVisibility } = this.state
 		const { classes, t } = this.props
-		console.log(this.props)
 		return <Fragment>
 			<ItemG>
 				<Hidden smDown>
